auth/requirements.txt
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
Deprecated==1.2.13
flasgger==0.9.5
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
mistune==2.0.2
packaging==21.3
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.8
pyrsistent==0.18.1
PyYAML==6.0
redis==4.2.2
six==1.16.0
SQLAlchemy==1.4.35
typing_extensions==4.2.0
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
pyjwt==2.3.0
